requirements.txt
pandas==2.0.3
pytesseract==0.3.10
opencv-python==4.8.0
pyyaml==6.0
openai==1.30.0
